# mes-root/docs/mes-auto_swagger.yaml
openapi: 3.0.0
...
servers:
  - description: DEV
    url: https://localhost:9998
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/hengyi/mes-auto/1.0.0
info:
  description: auto api
  version: "1.0.0"
  title: Mes Auto
  contact:
    email: ixtf1984@gmail.com
...
paths:
...
  /api/AutoDoffingAdaptCheckSilks:
    post:
      tags:
        - api
      summary: 自动落筒-验证标签
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - silkCar
                - workshop
              properties:
                silkCar:
                  $ref: '#/components/schemas/EntityByCodeDTO'
                workshop:
                  $ref: '#/components/schemas/EntityDTO'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CheckSilk'
        '400':
          $ref: '#/components/responses/JException'
  /api/AutoDoffingAdaptEvents:
    post:
      tags:
        - api
      summary: 自动落筒
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AutoDoffingAdaptEvent'
      responses:
        '200':
          description: ok
          headers:
            X-JAPP-MES-IF-SIGN:
              schema:
                type: string
                format: JWT
              description: 接口调用日志，出错核对，请保存
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventHandleResultDTO'
        '400':
          $ref: '#/components/responses/JException'
  /api/ManualDoffingCheckSilks:
    post:
      tags:
        - api
      summary: 手动落筒-标签验证
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - silkCar
                - lineMachineCount
              properties:
                silkCar:
                  $ref: '#/components/schemas/EntityByCodeDTO'
                lineMachineCount:
                  type: number
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CheckSilk'
        '400':
          $ref: '#/components/responses/JException'
  /api/ManualDoffingEvents:
    post:
      tags:
        - api
      summary: 手动落筒
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManualDoffingEvent'
      responses:
        '200':
          description: ok
        '400':
          $ref: '#/components/responses/JException'
  /api/DyeingSampleDoffingCheckSilks:
    post:
      tags:
        - api
      summary: 标样丝落筒-标签验证
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - silkCar
                - workshop
              properties:
                silkCar:
                  $ref: '#/components/schemas/EntityByCodeDTO'
                workshop:
                  $ref: '#/components/schemas/EntityDTO'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CheckSilk'
        '400':
          $ref: '#/components/responses/JException'
  /api/DyeingSampleDoffingEvents:
    post:
      tags:
        - api
      summary: 标样丝落筒
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AutoDoffingAdaptEvent'
      responses:
        '200':
          description: ok
        '400':
          $ref: '#/components/responses/JException'
  /api/PhysicalInfoDoffingCheckSilks:
    post:
      tags:
        - api
      summary: 取样丝落筒-标签验证
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - silkCar
                - workshop
              properties:
                silkCar:
                  $ref: '#/components/schemas/EntityByCodeDTO'
                workshop:
                  $ref: '#/components/schemas/EntityDTO'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CheckSilk'
        '400':
          $ref: '#/components/responses/JException'
  /api/PhysicalInfoDoffingEvents:
    post:
      tags:
        - api
      summary: 取样丝落筒
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AutoDoffingAdaptEvent'
      responses:
        '200':
          description: ok
        '400':
          $ref: '#/components/responses/JException'
  /api/SilkCarRuntimeDeleteEvents:
    post:
      tags:
        - api
      summary: 删除落筒
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - silkCarRecord
              properties:
                silkCarRecord:
                  $ref: '#/components/schemas/SilkCarRecordDTO'
      responses:
        '200':
          description: ok
        '400':
          $ref: '#/components/responses/JException'
  /api/AppendDoffingCheckSilks:
    post:
      tags:
        - api
      summary: 生产班拼车-位与车-标签验证
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - silkCarRecord
                - lineMachineCount
              properties:
                silkCarRecord:
                  $ref: '#/components/schemas/SilkCarRecordDTO'
                lineMachineCount:
                  type: number
      responses:
        '200':
          description: ok
        '400':
          $ref: '#/components/responses/JException'
  /api/AppendDoffingEvents:
    post:
      tags:
        - api
      summary: 生产班拼车-位与车
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppendDoffingEvent'
      responses:
        '200':
          description: ok
        '400':
          $ref: '#/components/responses/JException'
  /api/SilkCarRuntimeFlipEvents:
    post:
      tags:
        - api
      summary: AB面翻转
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - silkCarRecord
              properties:
                silkCarRecord:
                  $ref: '#/components/schemas/SilkCarRecordDTO'
      responses:
        '200':
          description: ok
        '400':
          $ref: '#/components/responses/JException'
  /api/SilkNoteFeedbackEvents:
    post:
      tags:
        - api
      summary: 备注处理
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - silkCarRecord
                - silkNote
              properties:
                silkCarRecord:
                  $ref: '#/components/schemas/SilkCarRecordDTO'
                silkNote:
                  $ref: '#/components/schemas/EntityDTO'
      responses:
        '200':
          description: ok
        '400':
          $ref: '#/components/responses/JException'
  /api/CarpoolEvents:
    post:
      tags:
        - api
      summary: 拼车
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CarpoolEvent'
      responses:
        '200':
          description: ok
        '400':
          $ref: '#/components/responses/JException'
  /api/SilkCarRuntimeGradeEvents:
    post:
      tags:
        - api
      summary: 丝车等级
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SilkCarRuntimeGradeEvent'
      responses:
        '200':
          description: ok
        '400':
          $ref: '#/components/responses/JException'
  /api/silkCarRuntimes/{code}:
    get:
      tags:
        - api
      summary: 获取丝车信息
      parameters:
        - in: path
          name: code
          required: true
          description: 丝车编号
          schema:
            type: string
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SilkCarRuntime'
        '400':
          $ref: '#/components/responses/JException'
  /api/silkCarRuntimes/{code}/physicalInfo:
    get:
      tags:
        - api
      summary: 丝车送物检查信息
      parameters:
        - in: path
          name: code
          required: true
          description: 丝车编号
          schema:
            type: string
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SilkCarRuntimePhysicalInfoDTO'
        '400':
          $ref: '#/components/responses/JException'
  /api/DyeingSampleSubmitEvents:
    post:
      tags:
        - api
      summary: 标样丝提交
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DyeingSampleSubmitEvent'
      responses:
        '200':
          description: ok
        '400':
          $ref: '#/components/responses/JException'
  /api/SilkRuntimeDetachEvents:
    post:
      tags:
        - api
      summary: 丝锭解码
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SilkRuntimeDetachEvent'
      responses:
        '200':
          description: ok
        '400':
          $ref: '#/components/responses/JException'
  /api/DyeingPrepareFirstSubmitEvents:
    post:
      tags:
        - api
      summary: 一次织袜，正常抽样
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DyeingPrepareFirstSubmitEvent'
      responses:
        '200':
          description: ok
        '400':
          $ref: '#/components/responses/JException'
  /api/DyeingPrepareCrossSpindleSubmitEvents:
    post:
      tags:
        - api
      summary: 一次织袜，位与锭交织
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DyeingPrepareFirstSubmitEvent'
      responses:
        '200':
          description: ok
        '400':
          $ref: '#/components/responses/JException'
  /api/DyeingPrepareCrossLineMachineSubmitEvents:
    post:
      tags:
        - api
      summary: 一次织袜，位与位交织
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DyeingPrepareCrossLineMachineSubmitEvent'
      responses:
        '200':
          description: ok
        '400':
          $ref: '#/components/responses/JException'
  /api/DyeingPrepareMultiSubmitEvents:
    post:
      tags:
        - api
      summary: 多次织袜
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DyeingPrepareMultiSubmitEvent'
      responses:
        '200':
          description: ok
        '400':
          $ref: '#/components/responses/JException'
  /api/SilkCarRuntimeGradeSubmitEvents:
    post:
      tags:
        - api
      summary: 等级提交
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SilkCarRuntimeGradeSubmitEvent'
      responses:
        '200':
          description: ok
        '400':
          $ref: '#/components/responses/JException'
...
  /api/ManualPackageBoxEventsSimple:
    post:
      tags:
        - api
      summary: 人工打包
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManualPackageBoxEventsSimple'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackageBox'
        '400':
          $ref: '#/components/responses/JException'
  /api/ManualPackageBoxEvents:
    post:
      tags:
        - api
      summary: 人工打包--@Deprecated
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManualPackageBoxEvent'
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PackageBox'
        '400':
          $ref: '#/components/responses/JException'
  /api/TemporaryBoxIncrementEvents:
    post:
      tags:
        - api
      summary: 增加暂存箱
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemporaryBoxIncrementEvent'
      responses:
        '200':
          description: ok
        '400':
          $ref: '#/components/responses/JException'
...
components:
  securitySchemes:
    # arbitrary name for the security scheme
    bearerAuth:
      type: http
      scheme: bearer
      # optional, arbitrary value for documentation purposes
      bearerFormat: JWT
  responses:
...
    JException:
      description: |
        非200的异常一般都会处理过，没处理的都是http status默认

        如：401没认证，403没权限，404找不到，等等
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/JException'
  schemas:
    JException:
      type: object
      properties:
        errorCode:
          type: string
          example: E00000
        errorMessage:
          type: string
...
    EntityDTO:
      type: object
      required:
        - id
      properties:
        id:
          type: string
    EntityByCodeDTO:
      type: object
      required:
        - code
      properties:
        code:
          type: string
    CheckSilk:
      allOf:
        - $ref: '#/components/schemas/BasicSilkPosition'
        - type: object
          required:
            - code
          properties:
            code:
              type: string
    EventHandleResultDTO:
      type: object
      description: 事件源，可当接口日志
      properties:
        silkCarRecordId:
          type: string
          description: 丝车车次ID
        eventId:
          type: string
          description: 事件源ID
    SilkCarRecordDTO:
      allOf:
        - $ref: '#/components/schemas/EntityDTO'
        - type: object
          required:
            - silkCar
          properties:
            silkCar:
              $ref: '#/components/schemas/EntityDTO'
...
    SilkCarRuntimePhysicalInfoDTO:
      type: object
      properties:
        barCode:
          type: string
          description: 丝车条码
        batchNumber:
          type: string
          description: 批号
        spec:
          type: string
          description: 规格
        sampler:
          type: string
          description: 取样人
        product:
          type: string
          description: 产品
        silks:
          type: array
          items:
            type: object
            properties:
              productLine:
                type: string
                description: 产线
              item:
                type: string
                description: 位号
              fallTime:
                type: string
                description: 落次
              spindleNumber:
                type: string
                description: 丝锭条码
              abnormalCauses:
                type: string
                description: 丝锭异常信息
...
    SilkCarRuntime:
      type: object
      properties:
        silkCarRecord:
          $ref: '#/components/schemas/SilkCarRecord'
        silkRuntimes:
          type: array
          items:
            $ref: '#/components/schemas/SilkRuntime'
    AppendDoffingEvent:
      type: object
      required:
        - silkCar
        - lineMachineCount
        - checkSilks
      properties:
        silkCarRecord:
          $ref: '#/components/schemas/SilkCarRecordDTO'
        lineMachineCount:
          type: number
        checkSilks:
          type: array
          items:
            $ref: '#/components/schemas/CheckSilk'
    AutoDoffingAdaptEvent:
      type: object
      required:
        - silkCar
        - workshop
        - grade
        - checkSilks
      properties:
        silkCar:
          $ref: '#/components/schemas/EntityByCodeDTO'
        workshop:
          $ref: '#/components/schemas/EntityDTO'
        grade:
          $ref: '#/components/schemas/EntityDTO'
        checkSilks:
          type: array
          items:
            $ref: '#/components/schemas/CheckSilk'
...
    CarpoolEvent:
      type: object
      required:
        - silkCar
        - silks
      properties:
        silkCar:
          $ref: '#/components/schemas/EntityByCodeDTO'
        silks:
          type: array
          items:
            $ref: '#/components/schemas/EntityByCodeDTO'
    ManualDoffingEvent:
      type: object
      required:
        - silkCar
        - lineMachineCount
        - grade
      properties:
        silkCar:
          $ref: '#/components/schemas/EntityByCodeDTO'
        lineMachineCount:
          type: number
        grade:
          $ref: '#/components/schemas/EntityDTO'
        checkSilks:
          type: array
          items:
            $ref: '#/components/schemas/CheckSilk'
...
    SilkCarRuntimeGradeEvent:
      type: object
      required:
        - silkCarRecord
        - grade
      properties:
        silkCarRecord:
          $ref: '#/components/schemas/SilkCarRecordDTO'
        grade:
          $ref: '#/components/schemas/EntityDTO'
...
    DyeingSampleSubmitEvent:
      type: object
      required:
        - silkCarRecord
        - silkRuntimes
      properties:
        silkCarRecord:
          $ref: '#/components/schemas/SilkCarRecordDTO'
        silkRuntimes:
          type: array
          items:
            $ref: '#/components/schemas/SilkRuntimeDTO'
    SilkRuntimeDetachEvent:
      type: object
      required:
        - silkCarRecord
        - silkRuntimes
      properties:
        silkCarRecord:
          $ref: '#/components/schemas/SilkCarRecordDTO'
        silkRuntimes:
          type: array
          items:
            $ref: '#/components/schemas/SilkRuntimeDTO'
    DyeingPrepareFirstSubmitEvent:
      type: object
      required:
        - silkCarRecord
        - silkRuntimes
        - dyeingSample
      properties:
        silkCarRecord:
          $ref: '#/components/schemas/SilkCarRecordDTO'
        silkRuntimes:
          type: array
          items:
            $ref: '#/components/schemas/SilkRuntimeDTO'
        dyeingSample:
          $ref: '#/components/schemas/EntityByCodeDTO'
    DyeingPrepareCrossLineMachineSubmitEvent:
      type: object
      required:
        - silkCarRecord1
        - silkRuntimes1
        - silkCarRecord2
        - silkRuntimes2
      properties:
        silkCarRecord1:
          $ref: '#/components/schemas/SilkCarRecordDTO'
        silkRuntimes1:
          type: array
          items:
            $ref: '#/components/schemas/SilkRuntimeDTO'
        silkCarRecord2:
          $ref: '#/components/schemas/SilkCarRecordDTO'
        silkRuntimes2:
          type: array
          items:
            $ref: '#/components/schemas/SilkRuntimeDTO'
    DyeingPrepareMultiSubmitEvent:
      type: object
      required:
        - dyeingType
        - silkCarRecord
        - silkRuntimes
        - dyeingSample
      properties:
        dyeingType:
          type: string
          enum: [SECOND, THIRD]
          description: 区分二次织袜或三次织袜
        silkCarRecord:
          $ref: '#/components/schemas/SilkCarRecordDTO'
        silkRuntimes:
          type: array
          items:
            $ref: '#/components/schemas/SilkRuntimeDTO'
        dyeingSample:
          $ref: '#/components/schemas/EntityByCodeDTO'
...
    SilkCarRuntimeGradeSubmitEvent:
      type: object
      required:
        - silkCarRecord
        - silkRuntimes
      properties:
        silkCarRecord:
          $ref: '#/components/schemas/SilkCarRecordDTO'
        silkRuntimes:
          type: array
          items:
            $ref: '#/components/schemas/SilkCarRuntimeGradeSubmitEvent.Silk'
    ManualPackageBoxEventsSimple:
      type: object
      required:
        - silkCount
        - silkCarRecords
      properties:
        silkCount:
          type: integer
        silkCarRecords:
          type: array
          items:
            $ref: '#/components/schemas/SilkCarRecordDTO'
...
    PackageBox:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
          description: 唛头条码
        palletCode:
          type: string
          description: 久鼎塑托
        type:
          type: string
          enum: [AUTO, MANUAL,FOREIGN]
          description: 打包类型，自动打包，人工打包，外贸打包
        grossWeight:
          type: number
          description: 毛重
        netWeight:
          type: number
          description: 净重
        batch:
          $ref: '#/components/schemas/Batch'
        printClass:
          description: 打印班次
          $ref: '#/components/schemas/PackageClass'
        printDate:
          type: number
          description: 打印日期
        budatClass:
          description: 入库班次
          $ref: '#/components/schemas/PackageClass'
        budat:
          type: number
          description: 入库日期
        silkCount:
          type: integer
          description: 丝锭颗数
        silks:
          type: array
          items:
            $ref: '#/components/schemas/Silk'
        createDateTime:
          type: number
          description: 打包时间
...
security:
  - bearerAuth: []